+++ src/r-c-ocaml-2-build_host.sh
@@ -46,7 +46,7 @@ usage() {
         printf "%s\n" "   -t DIR: Target directory for the reproducible directory tree"
         printf "%s\n" "   -b PREF: Required and used only for the MSVC compiler. See r-c-ocaml-1-setup.sh"
         printf "%s\n" "   -B Only build bytecode compiler and libraries."
-        printf "%s\n" "   -3 Use __ex32 ABI modifier so OCaml (not C) is 32-bit. Effective only with -B, and works on 64-bit targets."
+        printf "%s\n" "   -X Use bx32 ABI so OCaml (not C) is 32-bit. Effective only with -B, and works on 64-bit targets."
         printf "%s\n" "   -c OCAMLC_OPT_EXE: If a possibly older 'ocamlc.opt' is specified, it speeds up compilation of the new OCaml compiler"
         printf "%s\n" "   -e DKMLHOSTABI: Uses the DkML compiler detector find a host ABI compiler"
         printf "%s\n" "   -f HOSTSRC_SUBDIR: Use HOSTSRC_SUBDIR subdirectory of -t DIR to place the source code of the host ABI"
@@ -78,11 +78,11 @@ OCAMLC_OPT_EXE=
 FLEXLINKFLAGS=
 DISABLE_EXTRAS=0
 BYTECODEONLY=OFF
-BYTECODE32=OFF
+BX32=OFF
 SANITIZE_ADDRESS=OFF
 SANITIZE_LEAK=OFF
 export MSVS_PREFERENCE=
-while getopts ":s:d:t:B3b:c:e:m:k:l:rf:p:q:wALh" opt; do
+while getopts ":s:d:t:BXb:c:e:m:k:l:rf:p:q:wALh" opt; do
     case ${opt} in
         h )
             usage
@@ -103,7 +103,7 @@ while getopts ":s:d:t:B3b:c:e:m:k:l:rf:p:q:wALh" opt; do
             TARGETDIR="$OPTARG"
         ;;
         B ) BYTECODEONLY=ON ;;
-        3 ) BYTECODE32=ON ;;
+        X ) BX32=ON ;;
         b )
             MSVS_PREFERENCE="$OPTARG"
         ;;
@@ -251,7 +251,7 @@ log_trace ocaml_configure "$OCAMLHOST_UNIX" "$DKMLHOSTABI" \
     "$CONFIGUREARGS"
 
 # Extend m.h
-if [ "$BYTECODEONLY" = ON ] && [ "$BYTECODE32" = ON ]; then
+if [ "$BYTECODEONLY" = ON ] && [ "$BX32" = ON ]; then
     print_m_h_extensions "$DKMLHOSTABI" __ex32 >> runtime/caml/m.h
 else
     print_m_h_extensions "$DKMLHOSTABI" "" >> runtime/caml/m.h